requirements.txt
openpyxl==3.1.2
Pillow==10.1.0
unoconv==0.9.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.0
bcrypt==4.3.0
